[XLA] Introduce unrolling as a way to eliminate loop aliasing copies
Pipelined loops have inherent aliasing interference in them, due to loop inputs shift positions across iterations. This results in copy insertion adding copies for each pipelined input. In some cases extra copies on top of this are needed to properly sequence all the mandatory aliasing copies. Overall, this type of overhead makes pipelined loops a non-viable choice in HLO.

It is not necessary to insert copies to resolve interference in this case. The loop inputs, despite directly carried out as loop outputs, still have finite lifetimes across a certain amount of loop iterations. If the loop was unrolled just enough times to have the lifetimes of its inputs end before the outputs would be materialized, this would implicitly remove any sort of interference. The drawback of this approach is that it can in some cases drastically increase compile times due to linearly increasing graph size.

Add an option to CopyInsertion to perform this type of loop preprocessing before any copies would be inserted. Users can opt in if they want to trade compile time for execution time.

PiperOrigin-RevId: 671128753